Refine data query and subscription
