+++ nautilus_trader/data/messages.pyx
@@ -186,7 +186,8 @@ cdef class SubscribeData(DataCommand):
             The converted request message.
         """
         params = self.params.copy() if self.params else {}
-        params["subscription_name"] = f"{self.data_type.type.__name__}.{self.instrument_id}"
+        subscription_id = params.get("subscription_id") or self.instrument_id
+        params["subscription_name"] = f"{self.data_type.type.__name__}.{subscription_id}"
 
         return RequestData(
             data_type=self.data_type,
